+++ CharacterSettings.lua
@@ -3,6 +3,23 @@ local wticc = WrapTextInColorCode
 
 MIOG_CharacterSettings = {}
 
+function miog:SetCharacterSetting(key, value)
+    MIOG_CharacterSettings[key] = value
+
+end
+
+function miog:GetCharacterSetting(...)
+    local setting = MIOG_CharacterSettings
+
+    for k, v in ipairs({...}) do
+        setting = setting[v]
+
+    end
+
+    return setting
+
+end
+
 miog.characterSettings = {
     {name = "AppDialog texts", variableName = "MIOG_AppDialogTexts", key="appDialogTexts", default={}},
     {name = "App dialog box extended", variableName = "MIOG_AppDialogBoxExtented", key="appDialogBoxExtented", default=true},